+++ History.md
@@ -1,4 +1,9 @@
+unreleased
+==========
+
+  * Move to `jshttp` orgainzation
+
 0.0.0 / 2014-05-13
 ==================
 
-  * Genesis from `connect`
+  * Initial implementation